+++ tests/driver_srf02/main.c
@@ -75,7 +75,7 @@ static int cmd_sample(int argc, char **argv)
 {
     (void)argc;
     (void)argv;
-    uint32_t wakeup = xtimer_now();
+    uint64_t wakeup = xtimer_now();
 
     while(1) {
         sample();
